[Modify] 카테고리바 수정
